+++ src/lib/config.ts
@@ -62,8 +62,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "4.14.1-alpha.1758626754698-b3680673",
-  sdkVersion: "0.1.0-alpha.1",
-  genVersion: "2.716.10",
+  sdkVersion: "0.1.0-alpha.2",
+  genVersion: "2.716.16",
   userAgent:
-    "speakeasy-sdk/typescript 0.1.0-alpha.1 2.716.10 4.14.1-alpha.1758626754698-b3680673 cribl-control-plane",
+    "speakeasy-sdk/typescript 0.1.0-alpha.2 2.716.16 4.14.1-alpha.1758626754698-b3680673 cribl-control-plane",
 } as const;